+++ src/PluginFx.cpp
@@ -58,7 +58,7 @@ PluginFx::PluginFx() {
   Gain = 1.0;
 }
 
-void PluginFx::init(int sr) {
+void PluginFx::init(uint16_t sr) {
   mm = 0;
   s1 = s2 = s3 = s4 = c = d = 0;
   R24 = 0;
@@ -104,7 +104,7 @@ inline float PluginFx::NR(float sample, float g) {
   return y;
 }
 
-void PluginFx::process(float *work, int sampleSize) {
+void PluginFx::process(float *work, uint16_t sampleSize) {
   // very basic DC filter
   float t_fd = work[0];
   work[0] = work[0] - dc_id + dc_r * dc_od;
@@ -120,12 +120,12 @@ void PluginFx::process(float *work, int sampleSize) {
   // Gain
   if (Gain == 0.0)
   {
-    for (int i = 0; i < sampleSize; i++ )
+    for (uint16_t i = 0; i < sampleSize; i++ )
       work[i] = 0.0;
   }
   else if ( Gain != 1.0)
   {
-    for (int i = 0; i < sampleSize; i++ )
+    for (uint16_t i = 0; i < sampleSize; i++ )
       work[i] *= Gain;
   }
 
@@ -152,7 +152,7 @@ void PluginFx::process(float *work, int sampleSize) {
   float g = rCutoff;
   float lpc = g / (1 + g);
 
-  for (int i = 0; i < sampleSize; i++ ) {
+  for (uint16_t i = 0; i < sampleSize; i++ ) {
     float s = work[i];
     s = s - 0.45 * tptlpupw(c, s, 15, sampleRateInv);
     s = tptpc(d, s, bright);
