+++ src/dexed.h
@@ -51,7 +51,7 @@
 #define NUM_VOICE_PARAMETERS 156
 
 struct ProcessorVoice {
-  int16_t midi_note;
+  uint8_t midi_note;
   uint8_t velocity;
   int16_t porta;
   bool keydown;
@@ -165,7 +165,7 @@ enum ENGINES {
 class Dexed
 {
   public:
-    Dexed(uint8_t maxnotes, int rate);
+    Dexed(uint8_t maxnotes, uint16_t rate);
     ~Dexed();
 
     // Global methods
@@ -199,22 +199,22 @@ class Dexed
 #ifndef TEENSYDUINO
     void setCompressor(bool comp);
     bool getCompressor(void);
-    void setCompressorPreGain_dB(float32_t pre_gain);
-    void setCompressorAttack_sec(float32_t attack_sec);
-    void setCompressorRelease_sec(float32_t release_sec);
-    void setCompressorThresh_dBFS(float32_t thresh_dBFS);
-    void setCompressionRatio(float32_t comp_ratio);
-    float32_t getCompressorPreGain_dB(void);
-    float32_t getCompressorAttack_sec(void);
-    float32_t getCompressorRelease_sec(void);
-    float32_t getCompressorThresh_dBFS(void);
-    float32_t getCompressionRatio(void);
+    void setCompressorPreGain_dB(float pre_gain);
+    void setCompressorAttack_sec(float attack_sec);
+    void setCompressorRelease_sec(float release_sec);
+    void setCompressorThresh_dBFS(float thresh_dBFS);
+    void setCompressionRatio(float comp_ratio);
+    float getCompressorPreGain_dB(void);
+    float getCompressorAttack_sec(void);
+    float getCompressorRelease_sec(void);
+    float getCompressorThresh_dBFS(void);
+    float getCompressionRatio(void);
 #endif
     int16_t checkSystemExclusive(const uint8_t* sysex, const uint16_t len);
 
     // Sound methods
-    void keyup(int16_t pitch);
-    void keydown(int16_t pitch, uint8_t velo);
+    void keyup(uint8_t pitch);
+    void keydown(uint8_t pitch, uint8_t velo);
     void setSustain(bool sustain);
     bool getSustain(void);
     void panic(void);
@@ -355,12 +355,12 @@ class Dexed
       03, 48,                                                                             // pitch_mod_sensitivity, transpose
       73, 78, 73, 84, 32, 86, 79, 73, 67, 69                                              // 10 * char for name ("INIT VOICE")
     };
-    float32_t samplerate;
+    float samplerate;
     uint8_t data[NUM_VOICE_PARAMETERS];
     uint8_t max_notes;
     PluginFx fx;
     Controllers controllers;
-    int lastKeyDown;
+    int32_t lastKeyDown;
     uint32_t xrun;
     uint16_t render_time_max;
     int16_t currentNote;
@@ -375,9 +375,9 @@ class Dexed
     EngineMsfa* engineMsfa;
     EngineMkI* engineMkI;
     EngineOpl* engineOpl;
-    void getSamples(float32_t* buffer, uint16_t n_samples);
+    void getSamples(float* buffer, uint16_t n_samples);
     void getSamples(int16_t* buffer, uint16_t n_samples);
-    void compress(float32_t* wav_in, float32_t* wav_out, uint16_t n, float32_t threshold, float32_t slope, uint16_t sr,  float32_t tla, float32_t twnd, float32_t tatt, float32_t trel);
+    void compress(float* wav_in, float* wav_out, uint16_t n, float threshold, float slope, uint16_t sr,  float tla, float twnd, float tatt, float trel);
     bool use_compressor;
     uint8_t velocity_offset;
     uint8_t velocity_max;
